requirements.txt
selenium==4.16.0
pillow==10.1.0
ddddocr==1.4.7
requests==2.31.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
webdriver_manager==4.0.1
websockets==12.0
aiohttp==3.9.1
pymysql==1.1.0
pycryptodome==3.20.0